added global exception handling, cascade, fixed bugs

+++ Online-Medicine-Shopping-System/src/main/java/com/capg/controller/AdminController.java
@@ -15,6 +15,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.capg.entity.Admin;
+import com.capg.exception.AdminAlreadyRegisteredException;
+import com.capg.exception.AdminNotFoundException;
 import com.capg.service.AdminService;
 
 
@@ -34,17 +36,27 @@ public class AdminController {
 	@GetMapping("/{id}")
 	public ResponseEntity<Admin> findAdminById(@PathVariable long id) {
 		Optional<Admin> admin = adminService.findAdminById(id);
+		if(admin.isEmpty()) {
+			throw new AdminNotFoundException("No Admin found with the given id: " + id);
+		}
 		return ResponseEntity.ok(admin.get());
 	}
 
 	@PostMapping("/save")
 	public ResponseEntity<Admin> saveAdmin(@RequestBody Admin admin) {
+		if(!adminService.getAllAdmins().isEmpty()) {
+			throw new AdminAlreadyRegisteredException("An Admin is already registered. Please login!");
+		}
 		Admin savedAdmin = adminService.saveAdmin(admin);
 		return ResponseEntity.ok(savedAdmin);
 	}
 
 	@DeleteMapping("/delete/{id}")
 	public void deleteByid(@PathVariable long id) {
+		Optional<Admin> admin = adminService.findAdminById(id);
+		if(admin.isEmpty()) {
+			throw new AdminNotFoundException("No Admin found with the given id: " + id);
+		}
 		adminService.deleteAdminById(id);
 	}
 }
