+++ Online-Medicine-Shopping-System/src/main/java/com/capg/controller/CustomerController.java
@@ -15,8 +15,12 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.capg.entity.Customer;
+import com.capg.entity.Order;
+import com.capg.exception.CustomerNotFoundException;
+import com.capg.exception.NoCustomerPresentException;
+import com.capg.exception.OrderNotFoundException;
 import com.capg.service.CustomerService;
-
+import com.capg.service.OrderService;
 
 @CrossOrigin(origins = "http://localhost:3000")
 @RestController
@@ -24,16 +28,24 @@ import com.capg.service.CustomerService;
 public class CustomerController {
 	@Autowired
 	private CustomerService customerService;
+	@Autowired
+	private OrderService orderService;
 
 	@GetMapping("")
 	public ResponseEntity<List<Customer>> getAllCustomers() {
 		List<Customer> list = customerService.getAllCustomers();
+		if (list.isEmpty()) {
+			throw new NoCustomerPresentException("There are no customers present in the database!");
+		}
 		return ResponseEntity.ok(list);
 	}
 
 	@GetMapping("/{id}")
 	public ResponseEntity<Customer> findCustomerById(@PathVariable long id) {
 		Optional<Customer> customer = customerService.findCustomerById(id);
+		if (customer.isEmpty()) {
+			throw new CustomerNotFoundException("No customer found with id: " + id);
+		}
 		return ResponseEntity.ok(customer.get());
 	}
 
@@ -45,7 +57,20 @@ public class CustomerController {
 
 	@DeleteMapping("/delete/{id}")
 	public void deleteByid(@PathVariable long id) {
+		Optional<Customer> customer = customerService.findCustomerById(id);
+		if (customer.isEmpty()) {
+			throw new CustomerNotFoundException("No customer found with id: " + id);
+		}
 		customerService.deleteCustomerById(id);
 	}
-}
 
+	@GetMapping("/of-order/{orderId}")
+	public ResponseEntity<Customer> getCustomerByOrderId(@PathVariable int orderId) {
+		Optional<Order> order = orderService.findOrderById(orderId);
+		if (order.isEmpty()) {
+			throw new OrderNotFoundException("No Order found with the given id: " + orderId);
+		}
+		Customer customer = order.get().getCustomer();
+		return ResponseEntity.ok(customer);
+	}
+}
