+++ Online-Medicine-Shopping-System/src/main/java/com/capg/controller/MedicineController.java
@@ -17,8 +17,14 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.capg.entity.Category;
 import com.capg.entity.Medicine;
+import com.capg.entity.Order;
+import com.capg.exception.CategoryNotFoundException;
+import com.capg.exception.MedicineNotFoundException;
+import com.capg.exception.NoMedicinePresentException;
+import com.capg.exception.OrderNotFoundException;
 import com.capg.service.CategoryService;
 import com.capg.service.MedicineService;
+import com.capg.service.OrderService;
 
 @CrossOrigin(origins = "http://localhost:3000")
 @RestController
@@ -26,19 +32,26 @@ import com.capg.service.MedicineService;
 public class MedicineController {
 	@Autowired
 	private MedicineService medicineService;
-
+	@Autowired
+	private OrderService orderService;
 	@Autowired
 	private CategoryService categoryService;
 
 	@GetMapping("")
 	public ResponseEntity<List<Medicine>> getAllMedicines() {
 		List<Medicine> list = medicineService.getAllMedicines();
+		if (list.isEmpty()) {
+			throw new NoMedicinePresentException("There are no medicines present in the database!");
+		}
 		return ResponseEntity.ok(list);
 	}
 
 	@GetMapping("/{id}")
 	public ResponseEntity<Medicine> findMedicineById(@PathVariable long id) {
 		Optional<Medicine> medicine = medicineService.findMedicineById(id);
+		if (medicine.isEmpty()) {
+			throw new MedicineNotFoundException("No medicine found with id: " + id);
+		}
 		return ResponseEntity.ok(medicine.get());
 	}
 
@@ -50,15 +63,37 @@ public class MedicineController {
 
 	@DeleteMapping("/delete/{id}")
 	public void deleteByid(@PathVariable long id) {
+		Optional<Medicine> medicine = medicineService.findMedicineById(id);
+		if (medicine.isEmpty()) {
+			throw new MedicineNotFoundException("No medicine found with id: " + id);
+		}
 		medicineService.deleteMedicineById(id);
 	}
 
 	@PostMapping("/{medicineId}/assign-category/{categoryId}")
 	private ResponseEntity<Medicine> assignCategoryToMedicine(@PathVariable int medicineId,
 			@PathVariable int categoryId) {
-		Medicine medicine = medicineService.findMedicineById(medicineId).get();
-		Category category = categoryService.findCategoryById(categoryId).get();
+		Optional<Medicine> medicineOptional = medicineService.findMedicineById(medicineId);
+		if (medicineOptional.isEmpty()) {
+			throw new MedicineNotFoundException("No medicine found with id: " + medicineId);
+		}
+		Optional<Category> categoryOptional = categoryService.findCategoryById(categoryId);
+		if (categoryOptional.isEmpty()) {
+			throw new CategoryNotFoundException("No category found with id: " + categoryId);
+		}
+		Medicine medicine = medicineOptional.get();
+		Category category = categoryOptional.get();
 		medicine.setCategory(category);
 		return new ResponseEntity<Medicine>(medicineService.saveMedicine(medicine), HttpStatus.OK);
+	}
+
+	@GetMapping("/of-order/{orderId}")
+	public ResponseEntity<Medicine> getMedicineByOrderId(@PathVariable int orderId) {
+		Optional<Order> order = orderService.findOrderById(orderId);
+		if (order.isEmpty()) {
+			throw new OrderNotFoundException("No Order found with the given id: " + orderId);
+		}
+		Medicine medicine = order.get().getMedicine();
+		return ResponseEntity.ok(medicine);
 	}
 }
